@@ -2,7 +2,7 @@ package security_test
 
 import (
 	"fmt"
-	"io/ioutil"
+	"os"
 	"strconv"
 	"testing"
 
@@ -14,7 +14,7 @@ import (
 )
 
 func testAccResourceSecuritySAML() (*security.SAML, error) {
-	dat, err := ioutil.ReadFile("../../../examples/saml-testconfig.xml")
+	dat, err := os.ReadFile("../../../examples/saml-testconfig.xml")
 	if err != nil {
 		return nil, err
 	}
