@@ -12,8 +12,10 @@ import (
 
 func ResourcePrivilege() *schema.Resource {
 	return &schema.Resource{
-		Description: "Use this resource to create a Nexus privilege.",
+		DeprecationMessage: "This resource is deprecated. Please use the resource nexus_privilege_* instead.",
+		Description: `!> This resource is deprecated. Please use the resource "nexus_privilege_*" instead.
 
+Use this resource to create a Nexus privilege.`,
 		Create: resourcePrivilegeCreate,
 		Read:   resourcePrivilegeRead,
 		Update: resourcePrivilegeUpdate,
@@ -140,7 +142,7 @@ func resourcePrivilegeCreate(d *schema.ResourceData, m interface{}) error {
 
 	privilege := getPrivilegeFromResourceData(d)
 
-	if err := client.Security.Privilege.Create(privilege); err != nil {
+	if err := client.Deprecated.Privilege.Create(privilege); err != nil {
 		return err
 	}
 
@@ -169,7 +171,7 @@ func resourcePrivilegeUpdate(d *schema.ResourceData, m interface{}) error {
 	client := m.(*nexus.NexusClient)
 
 	privilege := getPrivilegeFromResourceData(d)
-	if err := client.Security.Privilege.Update(d.Id(), privilege); err != nil {
+	if err := client.Deprecated.Privilege.Update(d.Id(), privilege); err != nil {
 		return err
 	}
 
